pyproject.toml
[tool.poetry]
name = "hackathon_packager"
version = "0.1.0"
description = ""
authors = ["Yi-Jyun Pan <pan93412@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.4.1"
requests = "^2.25.1"
httpx = "^0.18.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
